add entry point for quetz jobs
